\name{rallies}
\docType{data}
\alias{rallies}
\title{Tea Party Rallies in Florida Counties}
\description{
This data set gives Census, voting, religion, and Tea Party organization and rallies for 67 counties in Florida.
}
\usage{rallies}
\format{A matrix containing 67 observations and 13 variables.}
\source{Subset of data created by Rory McVeigh, Kraig Beyerlein, Burrel Vann Jr., and Priyamvada Trivedi}
\references{
McVeigh, Rory, Kraig Beyerlein, Burrel Vann Jr., and Priyamvada Trivedi. ``Educational Segregation, Tea Party Organizations, and Battles over Distributive Justice.'' \emph{American Sociological Review 79}: 630-652.
}
\keyword{datasets}